refactor hotel update command

+++ src/Api/Dtos/HotelDtos/CreateHotelCommandDto.cs
@@ -4,7 +4,7 @@ public record CreateHotelCommandDto
 {
     public required string Name { get; init; }
     public required string Address { get; init; }
-    public float Rating { get; init; }
+    public decimal Rating { get; init; }
     public Guid? ManagerId { get; init; }
     public IEnumerable<Guid> RoomIds { get; init; } = [];
 }
