+++ src/Application/Hotels/Commands/CreateHotelCommand.cs
@@ -9,7 +9,7 @@ public record CreateHotelCommand
 {
     public required string Name { get; init; }
     public required string Address { get; init; }
-    public float Rating { get; init; }
+    public decimal Rating { get; init; }
     public Guid? ManagerId { get; init; }
     public IEnumerable<Guid> RoomIds { get; init; } = [];
 }
